BB-190: Split AliasList and AliasRow and rename appropriately
