+++ src/client/components/forms/creator.jsx
@@ -24,7 +24,7 @@ const request = require('superagent-bluebird-promise');
 const Nav = require('react-bootstrap').Nav;
 const NavItem = require('react-bootstrap').NavItem;
 
-const Aliases = require('./parts/aliases.jsx');
+const Aliases = require('./parts/alias-list.jsx');
 const CreatorData = require('./parts/creator-data.jsx');
 const LoadingSpinner = require('../loading-spinner.jsx');
 const RevisionNote = require('./parts/revision-note.jsx');
